Exceptions renamed to avoid aliases
